move getMBeanAttributeValues to common
