+++ src/main/java/com/linkedin/kmf/common/MbeanAttributeValue.java
@@ -7,7 +7,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.linkedin.kmf.jmx;
+package com.linkedin.kmf.common;
 
 public class MbeanAttributeValue {
   private final String _mbean;
